+++ clinica/src/main/java/com/dh/clinica/service/impl/PacienteService.java
@@ -1,9 +1,11 @@
 package com.dh.clinica.service.impl;
 
 import com.dh.clinica.entity.Paciente;
+import com.dh.clinica.exception.ResourceNotFoundException;
 import com.dh.clinica.repository.IPacienteRepository;
 import com.dh.clinica.service.IPacienteService;
 import org.springframework.stereotype.Service;
+import org.springframework.web.client.ResourceAccessException;
 
 import java.util.List;
 import java.util.Optional;
@@ -38,7 +40,14 @@ public class PacienteService implements IPacienteService {
 
     @Override
     public void eliminarPaciente(Integer id) {
-        pacienteRepository.deleteById(id);
+        Optional<Paciente> pacienteEncontrado = pacienteRepository.findById(id);
+        if(pacienteEncontrado.isPresent()){
+            pacienteRepository.deleteById(id);
+
+        }else{
+            throw new ResourceNotFoundException("El paciente "+ id +" no fue encontrado");
+        }
+
     }
 
     @Override
